requirements.txt
# requirements.txt for Pi Scanner (Flask Web UI)
# ──────────────────────────────────────────────────────
# Core Python packages
Flask>=2.2

# Optional: USB LED control (install based on your LED type)
# For serial USB devices:
# pyserial>=3.5
# For GPIO-connected LEDs:
# RPi.GPIO>=0.7.0

# System packages (apt-get) needed on Raspberry Pi:
# sudo apt-get install -y python3-flask fswebcam img2pdf tesseract-ocr ocrmypdf imagemagick avahi-daemon
# Optional for USB hub port control:
# sudo apt-get install -y uhubctl

# ML Pipeline dependencies
# ──────────────────────────────────────────────────────
# PyTorch (install based on your system - see https://pytorch.org/)
# For CPU:
# torch>=2.0.0
# torchvision>=0.15.0
# For GPU (CUDA 11.8):
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# Transformers for text models
transformers>=4.30.0
tokenizers>=0.13.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Progress bars and utilities
tqdm>=4.65.0
requests>=2.28.0

# OCR for text extraction from invoices
pytesseract>=0.3.10
# Note: Also need to install Tesseract OCR system package:
# sudo apt-get install -y tesseract-ocr

# Image processing for receipt detection and cropping
opencv-python>=4.8.0  # For receipt detection and image enhancement
